Moved app.js and removed folder
